+++ src/ruby/nodes/calls.js
@@ -136,40 +136,11 @@ function printMethodAddArg(path, opts, print) {
   return concat([methodDoc, argsDoc]);
 }
 
-// Sorbet type annotations look like the following:
-//
-//     {method_add_block
-//       [{method_add_arg
-//          [{fcall
-//             [{@ident "sig"}]},
-//           {args []}]},
-//        {brace_block [nil, {stmts}]}}]}
-//
-function isSorbetTypeAnnotation(node) {
-  const [callNode, blockNode] = node.body;
-
-  return (
-    callNode.type === "method_add_arg" &&
-    callNode.body[0].type === "fcall" &&
-    callNode.body[0].body[0].body === "sig" &&
-    callNode.body[1].type === "args" &&
-    callNode.body[1].body.length === 0 &&
-    blockNode
-  );
-}
-
 function printMethodAddBlock(path, opts, print) {
   const node = path.getValue();
 
   const [callNode, blockNode] = node.body;
   const [callDoc, blockDoc] = path.map(print, "body");
-
-  // Very special handling here for sorbet type annotations. They look like Ruby
-  // code, but they're not actually Ruby code, so we're not going to mess with
-  // them at all.
-  if (isSorbetTypeAnnotation(node)) {
-    return opts.originalText.slice(opts.locStart(node), opts.locEnd(node));
-  }
 
   // Don't bother trying to do any kind of fancy toProc transform if the option
   // is disabled.