+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - [@jeffcarbs], [@kddeisz] - Long strings with interpolated expressions should only break if the contents in the original source is broken.
 - [@johannesluedke], [@kddeisz] - Fix for rescues with inline comments.
 - [@johncsnyder], [@kddeisz] - Comments should not attach to nodes beyond a double newline.
+- [@blampe], [@kddeisz] - Comments inside of a broken method chain get dropped.
 
 ## [1.3.0] - 2021-01-05